+++ lib/puma/cluster.rb
@@ -332,6 +332,13 @@ module Puma
       # This is aligned with the output from Runner, see Runner#output_header
       log "*      Workers: #{@options[:workers]}"
 
+      if @options[:workers] == 1
+        log "! WARNING: Detected running cluster mode with 1 worker:"
+        log "! Running Puma in cluster mode with a single worker is often a misconfiguration."
+        log "! Consider running Puma in single-mode in order to reduce memory overhead."
+        log "! See: https://github.com/puma/puma/issues/2534"
+      end
+
       # Threads explicitly marked as fork safe will be ignored.
       # Used in Rails, but may be used by anyone.
       before = Thread.list.reject { |t| t.thread_variable_get(:fork_safe) }